requirements.txt
tomli >= 2.2.1
tomli-w >= 1.2.0
jaimead7-pyutils >= 0.0.10
python-snap7 >= 2.0.2
